+++ src/algorithms/xorec_benchmark.hpp
@@ -18,7 +18,7 @@ explicit XorecBenchmark(const BenchmarkConfig& config) noexcept;
   bool check_for_corruption() const noexcept override;
 
 protected:
-  uint32_t m_num_total_blocks;
+  size_t m_num_total_blocks;
 
   // Data Buffers
   std::unique_ptr<uint8_t[]> m_data_buffer;  ///< Buffer for the original data we want to transmit
